fix
-add logic to first modal
-delete id-Route
-broke Review-modal
