--- src/pages/statisticsPage/statisticsPage.jsx
@@ -1,25 +1,0 @@
-import { useSelector } from 'react-redux';
-import { useLocation, useNavigate } from 'react-router-dom';
-import { useEffect } from 'react';
-import { PageAnimation } from '../../components/PageAnimation/PageAnimation';
-
-const StatisticsPage = () => {
-  const { isLoggedIn } = useSelector(state => state.auth);
-
-  const navigate = useNavigate();
-  const location = useLocation();
-
-  useEffect(() => {
-    !isLoggedIn && navigate((location.pathname = '/login'));
-  });
-
-  return (
-    isLoggedIn && (
-      <PageAnimation>
-        <h1>statistics</h1>
-      </PageAnimation>
-    )
-  );
-};
-
-export default StatisticsPage;
